refactor: move deque to new `src/collections` directory
