+++ src/collections/deque.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "common.h"
+#include "../common.h"
 
 typedef struct
 {
